requirements.txt
Flask==2.2.2
ldap3==2.9.1
mysql-connector-python==8.1.0
pycryptodomex==3.19.0
Werkzeug==2.2.3
